requirements.txt
PyGithub==2.1.1
python-box==7.1.1
ensure==1.0.3
joblib==1.3.2
numpy==1.26.1
scikit-learn==1.3.2
tqdm==4.66.1
librosa==0.10.1
keras==2.14.0
tensorflow==2.14.0
Flask==3.0.0
dvc==3.33.3